# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Managing Personal Banking Accounts API",
    "description": "This folder contains requests for the “Managing Personal Bank Accounts” API, which includes functionality for user authentication, currency management, user profile handling, messaging, bank identity management, bank account operations, card management, and payment processing. Each section facilitates essential banking operations, enabling users to efficiently manage their personal bank accounts and related activities. \n \n For more information, please refer to the [Postman API documentation](https://documenter.getpostman.com/view/34110793/2sAXxV4Upk), where also testing scripts are available for each request. \n \n To install it, click on 'Run in Postman' and choose the desired type of application. Then choose the workplace where you will install the collection and click 'Import'.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8082"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
  "paths": {
...
    "/api/v1/messages": {
      "get": {
        "tags": [
          "Message"
        ],
        "summary": "List",
        "description": "Get a list of messages.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      },
      "post": {
        "tags": [
          "Message"
        ],
        "summary": "Send",
        "description": "Sending a message to a specific user.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "receiver_email": "<String>",
                  "content": "<String>"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "202": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
    "/api/v1/{account_id}/payments/{bank_name}": {
      "get": {
        "tags": [
          "Payment"
        ],
        "summary": "List",
        "description": "Get a list of payments.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Account UUID"
          },
          {
            "name": "bank_name",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Bank name"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/api/v1/{account_id}/payments/{bank_name}/{id}": {
      "get": {
        "tags": [
          "Payment"
        ],
        "summary": "Payment by id",
        "description": "Get a specific payment.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Account UUID"
          },
          {
            "name": "bank_name",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Bank name"
          },
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Payment UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/api/v1/{acoount_id}/payments": {
      "post": {
        "tags": [
          "Payment"
        ],
        "summary": "Create bank transfer",
        "description": "Create bank transfer.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "recipient_number": "<String>",
                  "amount": "<Number>",
                  "description": "<String>",
                  "type": "BANK_TRANSFER"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "acoount_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Account UUID"
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    }
  }
}